Enable new TI in kprove
Move the option to enable new type inference to InnerParsingOptions so it's available in kprove as well.
